add NewTimer and After methods to Clock
add NewTimer and After methods to Clock

This allows for the testing of code that previously used time.After and time.Timer.

It correctly handles Timer.Reset and Timer.Stop. To do so, instead of using time.Timer directly, a wrapping clock.Timer struct is introduced. In the normal clock case, this wrapper sends all method calls to the time.Timer it contains and the channel C on it is the same as the underlying time.Timer.C.

In the testing case, the implementation uses the concepts of "sends" to model the Go runtime's ability to pass time.Time objects down the C channel. This must happen multiple times if the Timer was reset between changes to the fake Clock's time and is tricky to get right without the send concept. You can see the documentation on the fakeTimer and send structs for more.